Bumped Mongoid to 4.0.0
Due to the way gemspec files work I was unable to use a :git target in
them. This means I had to use the Gemfile instead. This may result in
having to manually build the gem. If there is a better way around this
please let me know.

As soon as mongoid 4.0.0 is in rubygems we can revert this change